Pool/requirements.txt
pyteal==0.20.1
algosdk==2.4.0
python-dotenv==1.0.0 